Added back end thru node.js on port 3000
