+++ info.json
@@ -1,0 +1,200 @@
+{
+  "success": true,
+  "center": {
+    "lat": 45.6878893427986,
+    "lng": 9.621842266249075
+  },
+  "results": [
+    {
+      "id": 50282,
+      "name": "Curno",
+      "fuels": [
+        {
+          "id": 73573225,
+          "price": 1.779,
+          "name": "Benzina",
+          "fuelId": 1,
+          "isSelf": true
+        },
+        {
+          "id": 73573226,
+          "price": 1.629,
+          "name": "Gasolio",
+          "fuelId": 2,
+          "isSelf": true
+        },
+        {
+          "id": 73573228,
+          "price": 1.192,
+          "name": "Metano",
+          "fuelId": 3,
+          "isSelf": false
+        },
+        {
+          "id": 73573227,
+          "price": 1.192,
+          "name": "Metano",
+          "fuelId": 3,
+          "isSelf": true
+        }
+      ],
+      "location": {
+        "lat": 45.68332426509254,
+        "lng": 9.617299983614926
+      },
+      "insertDate": "2024-05-29T07:15:13+02:00",
+      "address": "VIA FERMI ANGOLO VIA CURNASCO SNC 24035 - CURNO BG",
+      "brand": "Auchan"
+    },
+    {
+      "id": 47362,
+      "name": "AF PETROL SRL",
+      "fuels": [
+        {
+          "id": 73550933,
+          "price": 1.779,
+          "name": "Benzina",
+          "fuelId": 1,
+          "isSelf": true
+        },
+        {
+          "id": 73550932,
+          "price": 1.629,
+          "name": "Gasolio",
+          "fuelId": 2,
+          "isSelf": true
+        }
+      ],
+      "location": {
+        "lat": 45.692214000123,
+        "lng": 9.618496000123
+      },
+      "insertDate": "2024-05-29T00:13:15+02:00",
+      "address": "VIA BERGAMO  40 24035 - CURNO BG",
+      "brand": "PompeBianche"
+    },
+    {
+      "id": 11492,
+      "name": "CURNO-SS.BRIANTEA 342 VIA BERGAMO 21",
+      "fuels": [
+        {
+          "id": 73511879,
+          "price": 1.799,
+          "name": "Benzina",
+          "fuelId": 1,
+          "isSelf": true
+        },
+        {
+          "id": 73511880,
+          "price": 1.649,
+          "name": "Gasolio",
+          "fuelId": 2,
+          "isSelf": true
+        },
+        {
+          "id": 73511881,
+          "price": 1.949,
+          "name": "Hi-Q Diesel",
+          "fuelId": 6,
+          "isSelf": true
+        },
+        {
+          "id": 73511882,
+          "price": 2.099,
+          "name": "HiQ Perform+",
+          "fuelId": 28,
+          "isSelf": true
+        }
+      ],
+      "location": {
+        "lat": 45.69200117253395,
+        "lng": 9.62216897916801
+      },
+      "insertDate": "2024-05-28T07:05:44+02:00",
+      "address": "SS.BRIANTEA 342 VIA BERGAMO 21 24035 - CURNO BG",
+      "brand": "Q8"
+    },
+    {
+      "id": 8053,
+      "name": " co.rina",
+      "fuels": [
+        {
+          "id": 73508466,
+          "price": 2.037,
+          "name": "Benzina",
+          "fuelId": 1,
+          "isSelf": false
+        },
+        {
+          "id": 73508465,
+          "price": 1.819,
+          "name": "Benzina",
+          "fuelId": 1,
+          "isSelf": true
+        },
+        {
+          "id": 73508468,
+          "price": 1.887,
+          "name": "Gasolio",
+          "fuelId": 2,
+          "isSelf": false
+        },
+        {
+          "id": 73508467,
+          "price": 1.669,
+          "name": "Gasolio",
+          "fuelId": 2,
+          "isSelf": true
+        },
+        {
+          "id": 73508464,
+          "price": 2.157,
+          "name": "Hi-Q Diesel",
+          "fuelId": 6,
+          "isSelf": false
+        },
+        {
+          "id": 73508463,
+          "price": 1.999,
+          "name": "Hi-Q Diesel",
+          "fuelId": 6,
+          "isSelf": true
+        }
+      ],
+      "location": {
+        "lat": 45.69124802532839,
+        "lng": 9.626941680908203
+      },
+      "insertDate": "2024-05-28T06:52:19+02:00",
+      "address": "VIA BRIANTEA, KM 4  24129 - BERGAMO BG",
+      "brand": "Q8"
+    },
+    {
+      "id": 24759,
+      "name": "CURNO S.S. 342 BRIANTEA KM.4,885 -Q8",
+      "fuels": [
+        {
+          "id": 73501485,
+          "price": 1.779,
+          "name": "Benzina",
+          "fuelId": 1,
+          "isSelf": true
+        },
+        {
+          "id": 73501486,
+          "price": 1.629,
+          "name": "Gasolio",
+          "fuelId": 2,
+          "isSelf": true
+        }
+      ],
+      "location": {
+        "lat": 45.69245442050466,
+        "lng": 9.616742851589947
+      },
+      "insertDate": "2024-05-28T01:03:36+02:00",
+      "address": "S.S. 342 BRIANTEA KM.4,885  24035 - CURNO BG",
+      "brand": "Q8"
+    }
+  ]
+}
