--- sss/index.html
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Stazioni di Servizio</title>
-    <script src="https://unpkg.com/osrm@2.6.3/dist/6.min.js"></script>
-</head>
-<body>
-    <a href="/Capolavoro/pgbrtt.html">Go to index.html</a>
-    <table id="fuelStations">
-        <thead>
-            <tr>
-                <th>Nome</th>
-                <th>Prezzo</th>
-                <th>Distanza</th>
-                <th>Carburante</th>
-                <th>Brand</th>
-                <th>Self</th>
-            </tr>
-        </thead>
-        <tbody>
-            <!-- Qui verranno inserite dinamicamente le righe della tabella -->
-        </tbody>
-    </table>
-
-    <script>
-        // Funzione per caricare dinamicamente i dati nella tabella
-        function loadFuelStations() {
-            fetch('info.json')
-            .then(response => response.json())
-            .then(jsonData => {
-                var fuelStations = []; // Array per memorizzare le stazioni di servizio
-
-                // Loop attraverso le stazioni di servizio e popola l'array
-                jsonData.results.forEach(function(station) {
-                    station.fuels.forEach(function(fuel) {
-                        fuelStations.push({
-                            name: station.name,
-                            price: fuel.price,
-                            distance: calculateDistance(jsonData.center.lat, jsonData.center.lng, station.location.lat, station.location.lng),
-                            fuelType: fuel.name,
-                            brand: station.brand,
-                            selfService: fuel.isSelf? 'Yes' : 'No'
-                        });
-                    });
-                });
-
-                // Ordina le stazioni di servizio per nome
-                fuelStations.sort(function(a, b) {
-                    return a.name.localeCompare(b.name); // Ordina per nome
-                });
-
-                // Genera righe della tabella
-                var tableBody = document.querySelector('#fuelStations tbody');
-                fuelStations.forEach(function(station) {
-                    
-                    var row = '<tr>';
-                    row += '<td>' + station.name + '</td>';
-                    row += '<td>' + station.price.toFixed(3) + "€" + '</td>';
-                    row += '<td>' + station.distance.toFixed(2) + ' km' + '</td>';
-                    row += '<td>' + station.fuelType + '</td>';
-                    row += '<td>' + station.brand + '</td>';
-                    row += '<td>' + station.selfService + '</td>';
-                    row += '</tr>';
-                    tableBody.innerHTML += row;
-                });
-                
-            })
-            .catch(error => {
-                console.error('Errore nel caricamento dei dati:', error);
-                });
-        }
-
-
-
-        function toRadians(degrees) {
-            return degrees * Math.PI / 180;
-        }
-
-        function calculateDistance(userLat, userLon, stationLat, stationLon) {
-            const R = 6371; // Earth's radius in kilometers
-            const dLat = toRadians(stationLat - userLat);
-            const dLon = toRadians(stationLon - userLon);
-            const a =
-                Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-                Math.cos(toRadians(userLat)) * Math.cos(toRadians(stationLat)) *
-                Math.sin(dLon / 2) * Math.sin(dLon / 2)
-            ;
-            const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-            const distance = R * c;
-            return distance;
-        }
-        
-
-        /*
-        function calculateDistance(lat1, lon1, lat2, lon2) {
-            const osrm = new OSRM();
-            const options = {
-                coordinates: [[lon1, lat1], [lon2, lat2]],
-                annotations: 'distance'
-            };
-
-            osrm.route(options, function(err, result) {
-                if (!err) {
-                console.log(`Distance: ${result.routes[0].distance} meters`);
-                } else {
-                console.error('Error:', err);
-                }
-            });
-        }
-        */
-
-        // Carica le stazioni di servizio quando la pagina viene caricata
-        window.onload = function() {
-            loadFuelStations();
-        };
-    </script>
-
-</body>
-</html>
